third_party/blink/web_tests/http/tests/inspector-protocol/network/websocket/offline-expected.txt

Verifies that Network.emulateNetworkConditions affects web sockets when offline is enabled.

Running test: setUp

Running test: testNoSocketOpenWhenOffline
onerror onclose 

Running test: testNoMessageSentWhenOffline
onopen onmessage Offline

Running test: testNoMessageRecievedWhenOffline
onmessage Hit me
